+++ .idea/misc.xml
@@ -13,7 +13,9 @@
         <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/layout/bottom_sheet_contents.xml" value="0.22" />
         <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/layout/fragment_home.xml" value="0.1" />
         <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/layout/fragment_main.xml" value="0.1" />
+        <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/layout/fragment_search_bottom_sheet_dialog.xml" value="0.18" />
         <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/layout/fragment_shuffle.xml" value="0.1" />
+        <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/layout/item_shop_list.xml" value="0.1" />
         <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/menu/bottom_nav_menu.xml" value="0.1" />
         <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/menu/bottom_navigation_menu.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.2351851851851852" />
