+++ app/src/main/java/com/android/example/tanmen/Controller/Fragment/DetailFragment.kt
@@ -1,20 +1,19 @@
-package com.android.example.tanmen
+package com.android.example.tanmen.Controller.Fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.android.example.tanmen.databinding.FragmentShuffleBinding
+import com.android.example.tanmen.databinding.FragmentDetailBinding
 
-class ShuffleFragment : Fragment() {
-    private lateinit var binding: FragmentShuffleBinding
-
+class DetailFragment : Fragment() {
+    private lateinit var binding: FragmentDetailBinding
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentShuffleBinding.inflate(inflater, container, false)
+        binding = FragmentDetailBinding.inflate(inflater, container, false)
         return binding.root
     }
