+++ app/src/main/java/com/android/example/tanmen/Controller/Fragment/MainFragment.kt
@@ -1,0 +1,57 @@
+package com.android.example.tanmen.Controller.Fragment
+
+import android.os.Bundle
+import android.util.Log
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.core.os.bundleOf
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.setFragmentResultListener
+import androidx.recyclerview.widget.LinearLayoutManager
+import com.android.example.tanmen.Model.Shop
+import com.android.example.tanmen.View.ShopListAdapter
+import com.android.example.tanmen.databinding.FragmentMainBinding
+
+class MainFragment : Fragment() {
+    private lateinit var binding: FragmentMainBinding
+    var adapter: ShopListAdapter? = null
+
+    companion object {
+        const val REQ_KEY: String = "shop"
+        private const val ARG_SHOP: String = "shop"
+
+        fun createArgments(shop: Shop): Bundle {
+            return bundleOf(ARG_SHOP to shop)
+        }
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        binding = FragmentMainBinding.inflate(inflater, container, false)
+        setFragmentResultListener(REQ_KEY) { _, bundle ->
+            val shop = bundle.getSerializable(ARG_SHOP) as Shop
+
+            val shopList = mutableListOf(
+                Shop(shop.image, shop.name, shop.address, shop.hours)
+            )
+
+            val recyclerView = binding.shopList
+            val layoutManager = LinearLayoutManager(recyclerView.context)
+            adapter = ShopListAdapter(shopList)
+            recyclerView.layoutManager = layoutManager
+            recyclerView.adapter = adapter
+
+        }
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+
+
+    }
+}
