+++ app/src/main/java/com/android/example/tanmen/Controller/Fragment/ShuffleFragment.kt
@@ -1,19 +1,20 @@
-package com.android.example.tanmen
+package com.android.example.tanmen.Controller.Fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.android.example.tanmen.databinding.FragmentDetailBinding
+import com.android.example.tanmen.databinding.FragmentShuffleBinding
 
-class DetailFragment : Fragment() {
-    private lateinit var binding: FragmentDetailBinding
+class ShuffleFragment : Fragment() {
+    private lateinit var binding: FragmentShuffleBinding
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentDetailBinding.inflate(inflater, container, false)
+        binding = FragmentShuffleBinding.inflate(inflater, container, false)
         return binding.root
     }
