+++ app/src/main/java/com/android/example/tanmen/Model/Shop.kt
@@ -1,8 +1,10 @@
-package com.android.example.tanmen
+package com.android.example.tanmen.Model
+
+import java.io.Serializable
 
 data class Shop(
     var image: String?,
     var name: String,
     var address: String?,
     var hours: String?
-)
+): Serializable
