+++ app/src/main/java/com/android/example/tanmen/View/BottomNavigationPagerAdapter.kt
@@ -1,8 +1,10 @@
-package com.android.example.tanmen
+package com.android.example.tanmen.View
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import com.android.example.tanmen.Controller.Fragment.MainFragment
+import com.android.example.tanmen.Controller.Fragment.ShuffleFragment
 
 class BottomNavigationPagerAdapter(fm : FragmentActivity) : FragmentStateAdapter(fm) {
     override fun getItemCount(): Int {
