+++ app/src/main/res/layout/fragment_main.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".MainFragment">
+    tools:context=".Controller.Fragment.MainFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/shop_List"
