+++ app/src/main/res/layout/fragment_shuffle.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ShuffleFragment">
+    tools:context=".Controller.Fragment.ShuffleFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
